+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\index.tsx":"1","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\plugins\\fetch-plugin.ts":"2","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\plugins\\unpkg-path-plugin.ts":"3","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-editor.tsx":"4"},{"size":3243,"mtime":1689954374139,"results":"5","hashOfConfig":"6"},{"size":3250,"mtime":1689612354887,"results":"7","hashOfConfig":"6"},{"size":2095,"mtime":1689266132893,"results":"8","hashOfConfig":"6"},{"size":2340,"mtime":1689954374138,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1adxqzh",{"filePath":"12","messages":"13","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"16","messages":"17","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\index.tsx",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\plugins\\fetch-plugin.ts",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\plugins\\unpkg-path-plugin.ts",["18"],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-editor.tsx",[],{"ruleId":"19","severity":1,"message":"20","line":1,"column":8,"nodeType":"21","messageId":"22","endLine":1,"endColumn":13},"@typescript-eslint/no-unused-vars","'axios' is defined but never used.","Identifier","unusedVar"]
+[{"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\index.tsx":"1","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\plugins\\fetch-plugin.ts":"2","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\plugins\\unpkg-path-plugin.ts":"3","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-editor.tsx":"4","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\preview.tsx":"5"},{"size":2144,"mtime":1689955679928,"results":"6","hashOfConfig":"7"},{"size":3250,"mtime":1689612354887,"results":"8","hashOfConfig":"7"},{"size":2095,"mtime":1689266132893,"results":"9","hashOfConfig":"7"},{"size":2340,"mtime":1689954374138,"results":"10","hashOfConfig":"7"},{"size":1091,"mtime":1689955426207,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1adxqzh",{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"17","messages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"19","usedDeprecatedRules":"16"},{"filePath":"20","messages":"21","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"22","messages":"23","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\index.tsx",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\plugins\\fetch-plugin.ts",[],["24","25"],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\plugins\\unpkg-path-plugin.ts",["26"],"import axios from \"axios\";\r\nimport localforage from \"localforage\";\r\n\r\nimport * as esbuild from \"esbuild-wasm\";\r\n\r\nexport const fileCache = localforage.createInstance({\r\n  // configuration object\r\n  name: \"filecache\",\r\n});\r\n\r\nexport const unpkgPathPlugin = () => {\r\n  // return a object is a plugin that works inside of its esbuild.\r\n  /* \r\n  return an object\r\n    - name (string): for debugging\r\n    - setup (function):\r\n        + is going to be called automatically by esbuild with a single argument - build\r\n            + build argument: represents the building process: the entire process is: \r\n                + finding some file\r\n                + parsing it\r\n                + transpiling it\r\n                + joining a bunch of different files together\r\n            + we can override behaviour of build process with .onResolve, .onLoad\r\n\r\n  */\r\n  return {\r\n    name: \"unpkg-path-plugin\",\r\n    setup(build: esbuild.PluginBuild) {\r\n      /* \r\n      trying to figure out where the file is stored or what the actual path to the file  \r\n      */\r\n      //  Handle root entry file of index.js\r\n      build.onResolve({ filter: /(^index\\.js$)/ }, () => {\r\n        console.log(\"onResolve index.js\");\r\n        return {\r\n          path: \"index.js\",\r\n          namespace: \"a\",\r\n        };\r\n      });\r\n\r\n      // Handle relative paths in module\r\n      build.onResolve({ filter: /^\\.+\\// }, (args: any) => {\r\n        console.log(\"onResolve\", args);\r\n        return {\r\n          namespace: \"a\",\r\n          path: new URL(args.path, \"https://unpkg.com\" + args.resolveDir + \"/\")\r\n            ?.href,\r\n        };\r\n      });\r\n\r\n      // handle main file of module\r\n      build.onResolve({ filter: /.*/ }, async (args: any) => {\r\n        console.log(\"onResolve\", args);\r\n        return {\r\n          namespace: \"a\",\r\n          path: `https://unpkg.com/${args.path}`,\r\n        };\r\n        /* else if (args.path === \"tiny-test-pkg\") {\r\n          return {\r\n            path: \"https://unpkg.com/tiny-test-pkg@1.0.0/index.js\",\r\n            namespace: \"a\",\r\n          };\r\n        } */\r\n      });\r\n    },\r\n  };\r\n};\r\n","D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\code-editor.tsx",[],"D:\\Developments\\typescript\\learn-ts\\jbook\\src\\components\\preview.tsx",[],{"ruleId":"27","replacedBy":"28"},{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","severity":1,"message":"32","line":1,"column":8,"nodeType":"33","messageId":"34","endLine":1,"endColumn":13},"no-native-reassign",["35"],"no-negated-in-lhs",["36"],"@typescript-eslint/no-unused-vars","'axios' is defined but never used.","Identifier","unusedVar","no-global-assign","no-unsafe-negation"]
